rfdc: Use double type for Mixer settings frequency
This patch use double data type for Mixer settings
frequency instead of float type to support 48-bit
result.


Acked-for-series: Srinivas Goud <sgoud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/rfdc/src/xrfdc.h
@@ -175,7 +175,7 @@ typedef struct {
  * Mixer settings.
  */
 typedef struct {
-	float Freq;
+	double Freq;
 	float PhaseOffset;
 	u32 EventSource;
 	u32 FineMixerMode;
